requirement.txt
Flask
numpy
pandas
python-dotenv
mysql-connector_python
pymysql
#-e .
scikit-learn
sqlalchemy